PWN-5302 - SolendSDK: update naming and packages
