+++ app/src/main/java/org/p2p/wallet/AppModule.kt
@@ -35,7 +35,7 @@ import org.p2p.wallet.swap.SwapModule
 import org.p2p.wallet.transaction.di.TransactionModule
 import org.p2p.wallet.user.UserModule
 import org.p2p.wallet.user.repository.prices.di.TokenPricesModule
-import solendsdkfacade.di.P2PSdkModule
+import org.p2p.wallet.sdk.di.P2PSdkModule
 
 object AppModule {
     fun create(restartAction: () -> Unit) = module {
