+++ app/src/main/java/org/p2p/wallet/sdk/facade/SolendMethodResultHandler.kt
@@ -1,0 +1,9 @@
+package org.p2p.wallet.sdk.facade
+
+import org.p2p.wallet.sdk.facade.model.SolendMethodResultError
+import org.p2p.wallet.sdk.facade.model.SolendMethodResultSuccess
+
+interface SolendMethodResultHandler {
+    fun handleResultSuccess(result: SolendMethodResultSuccess)
+    fun handleResultError(error: SolendMethodResultError)
+}
