+++ app/src/main/java/org/p2p/wallet/sdk/facade/model/SolendEnvironment.kt
@@ -1,0 +1,5 @@
+package org.p2p.wallet.sdk.facade.model
+
+enum class SolendEnvironment(val sdkValue: String) {
+    DEVNET("devnet"), PRODUCTION("production")
+}
