+++ app/src/main/java/org/p2p/wallet/sdk/facade/model/SolendFeePayerTokenData.kt
@@ -1,0 +1,19 @@
+package org.p2p.wallet.sdk.facade.model
+
+import com.google.gson.annotations.SerializedName
+import org.p2p.wallet.utils.Base58String
+
+data class SolendFeePayerTokenData(
+    @SerializedName("sender_account_pubkey")
+    val senderAccountAddress: Base58String,
+    @SerializedName("recipient_account_pubkey")
+    val recipientAccountAddress: Base58String,
+    @SerializedName("mint_pubkey")
+    val tokenMintAddress: Base58String,
+    @SerializedName("authority_pubkey")
+    val transferAuthorityAddress: Base58String,
+    @SerializedName("exchange_rate")
+    val tokenExchangeRateLamports: Double,
+    @SerializedName("decimals")
+    val tokenDecimals: Int,
+)
