--- app/src/main/java/solendsdkfacade/SolendMethodResultHandler.kt
@@ -1,9 +1,0 @@
-package solendsdkfacade
-
-import solendsdkfacade.model.SolendMethodResultError
-import solendsdkfacade.model.SolendMethodResultSuccess
-
-interface SolendMethodResultHandler {
-    fun handleResultSuccess(result: SolendMethodResultSuccess)
-    fun handleResultError(error: SolendMethodResultError)
-}
