--- app/src/main/java/solendsdkfacade/model/SolendEnvironment.kt
@@ -1,5 +1,0 @@
-package solendsdkfacade.model
-
-enum class SolendEnvironment(val sdkValue: String) {
-    DEVNET("devnet"), PRODUCTION("production")
-}
