--- app/src/main/java/solendsdkfacade/model/SolendMethodResultError.kt
@@ -1,8 +1,0 @@
-package solendsdkfacade.model
-
-import com.google.gson.annotations.SerializedName
-
-data class SolendMethodResultError(
-    @SerializedName("error")
-    val error: String,
-) : Throwable(message = error)
