--- app/src/main/java/solendsdkfacade/model/SolendMethodResultSuccess.kt
@@ -1,11 +1,0 @@
-package solendsdkfacade.model
-
-sealed interface SolendMethodResultSuccess {
-    class SolendDepositTransactions(
-        val transactions: List<String>
-    ) : SolendMethodResultSuccess
-
-    class SolendWithdrawTransactions(
-        val transactions: List<String>
-    ) : SolendMethodResultSuccess
-}
